requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
Django==4.1.5
djangorestframework==3.14.0
exceptiongroup==1.1.0
filelock==3.9.0
frozenlist==1.3.3
h11==0.14.0
huggingface-hub==0.11.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
markdownify==0.11.6
multidict==6.0.4
numpy==1.24.1
openai==0.26.1
openapi==1.1.0
outcome==1.2.0
packaging==23.0
pyChatGPT==0.4.3.1
pycparser==2.21
pyrsistent==0.19.3
PySocks==1.7.1
pytz==2022.7.1
PyVirtualDisplay==3.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
undetected-chromedriver==3.2.1
urllib3==1.26.14
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
